Updating reservation feature completed and some bugs fixed

+++ components/reservation/Reservation.tsx
@@ -7,6 +7,7 @@ import { FaEdit } from 'react-icons/fa'
 import { MdDelete } from 'react-icons/md'
 import { useRouter } from 'next/router'
 import { dateFormat } from '../../utils/dateFormating'
+import Swal from 'sweetalert2'
 
 function Reservation({ reservation }) {
 
@@ -19,17 +20,47 @@ function Reservation({ reservation }) {
 
   function deleteHandler(reservationId:number, reservationAmount: number, reservationShow: number) {
     console.log('DELETE', reservationId, reservationAmount)
-    axios.delete('api/deleteReservation/', { data: { id: reservationId,
-      seats: reservationAmount,
-      show: reservationShow } }).
-      then((res) => {
-        console.log('SUCESSFULLY DELETED', res)
-      }).
-      catch((err) => console.log(err))
+    Swal.fire({
+      title: 'This reservation is being deleted',
+      icon: 'warning',
+      html: '<p>Are you sure that you want to delete this reservation?</p>',
+      confirmButtonText: 'Yes',
+      cancelButtonText: 'No',
+      showCancelButton: true
+    }).then((result) => {
+      if (result.isConfirmed) {
+        axios.delete('api/deleteReservation/', { data: { id: reservationId,
+          seats: reservationAmount,
+          show: reservationShow } }).
+          then((res) => {
+            Swal.fire({
+              title: 'Reservation successfully removed',
+              icon: 'success',
+              html: '<p>The reservation has been canceled. Let\'s check your other reservations.</p>',
+              confirmButtonText: 'OK'
+            }).then(() => {
+              // TODO: This reload is temporary the idea is to show the changes immediately
+              router.reload()
+            })
+          }).
+          catch((err) => {
+            console.log(err)
+            Swal.fire({
+              title: `${err.statusCode === undefined
+                ? 'Oops :('
+                : err.statusCode} Something went wrong`,
+              icon: 'error',
+              html: '<p>Please contact the admin and try again later</p>',
+              confirmButtonText: 'OK'
+            })
+          })
+      } else if (result.isDenied) {
+        Swal.fire('Delete reservetion has been canceled', '', 'info')
+      }
+    })
   }
 
-  function editHandler(reservationId:number, reservationAmount:number, reservationShow: number) {
-    console.log('EDIT', reservationId, reservationAmount, reservationShow)
+  function editHandler() {
     router.push({
       pathname: '/reservations/r',
       query: { data: JSON.stringify(reservation) }
@@ -37,7 +68,7 @@ function Reservation({ reservation }) {
   }
 
   return (
-    <article className="bg-gray-900 rounded-lg mx-2 my-4 py-3 text-center grid place-items-center grid-rows-3 grid-flow-col gap-2 md:w-2/6">
+    <article className="bg-gray-900 rounded-lg mx-2 my-4 p-3 text-center grid place-items-center grid-rows-3 grid-flow-col gap-2 md:w-2/6">
       <div className="row-span-3 grid place-items-center">
         <img alt="tickets" src={Tickets.src} width="50px" />
         <div className="grid place-items-center">
@@ -53,7 +84,7 @@ function Reservation({ reservation }) {
       <div className="col-span-2">
         <button
           className="inline-flex justify-center py-2 px-6 border border-transparent shadow-sm text-sm font-medium rounded-md text-white bg-fuchsia-600 hover:bg-fuchsia-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-fuchsia-500 mx-4"
-          onClick={() => editHandler(reservation.reservation_id, reservation.tickets_amount, reservation.show_id)}
+          onClick={() => editHandler(reservation)}
           type="button"
         ><FaEdit />
         </button>
